+++ src/main/java/com/nhnacademy/blog/category/dto/CategoryUpdateRequestDto.java
@@ -2,13 +2,13 @@ package com.nhnacademy.blog.category.dto;
 
 public class CategoryUpdateRequestDto {
     
-    private final Integer categoryId;
-    private final Integer categoryPid;
+    private final Long categoryId;
+    private final Long categoryPid;
     private final Integer topicId;
     private final String categoryName;
     private final int categorySec;
 
-    public CategoryUpdateRequestDto(Integer categoryId, Integer categoryPid, Integer topicId, String categoryName, int categorySec) {
+    public CategoryUpdateRequestDto(Long categoryId, Long categoryPid, Integer topicId, String categoryName, int categorySec) {
         this.categoryId = categoryId;
         this.categoryPid = categoryPid;
         this.topicId = topicId;
@@ -16,11 +16,11 @@ public class CategoryUpdateRequestDto {
         this.categorySec = categorySec;
     }
 
-    public Integer getCategoryId() {
+    public Long getCategoryId() {
         return categoryId;
     }
 
-    public Integer getCategoryPid() {
+    public Long getCategoryPid() {
         return categoryPid;
     }
